requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
django-tinymce==3.7.1
idna==3.6
pillow==10.2.0
python-gettext==5.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
django==5.0.2
django-resized
django-embed-video==1.4.10
django-allauth==0.61.1
django-crispy-forms==2.1
django-extensions==3.2.3
python-dotenv==1.0.1
django-extensions==3.2.3
django-rosetta==0.10.0
django_parler==2.3
django-redis
django-redis-sessions==0.6.2
django-environ==0.11.2
django-admin-honeypot
django-cookie-consent
django-allauth
django-extensions
django-crispy-forms

scipy
django-parler
django-cookie-consent
django-tinymce
django-resized
